Move check stream capturing to utils.py
